requirements.txt
fastapi
uvicorn
groq
playwright
jinja2
python-dotenv
pydantic
